+++ src/app/manifestGeneration/manifestGenerators/generator13/TabManifestUpdater.ts
@@ -7,10 +7,20 @@ import { GeneratorTeamsAppOptions } from "../../../GeneratorTeamsAppOptions";
 
 export class TabManifestUpdater implements IManifestUpdater {
     public updateManifest(manifest: any, options: GeneratorTeamsAppOptions): void {
-        (<any[]>manifest.configurableTabs).push({
-            configurationUrl: `https://{{HOSTNAME}}/${options.tabName}/config.html`,
-            canUpdateConfiguration: true,
-            scopes: options.tabScopes
-        });
+        if (options.tabType == "static") { 
+            (<any[]>manifest.staticTabs).push({
+                entityId: "default-data",
+                name: options.tabTitle,
+                contentUrl: `https://{{HOSTNAME}}/${options.tabName}/`,
+                scopes: ["personal"]
+            });
+        }
+        else {
+            (<any[]>manifest.configurableTabs).push({
+                configurationUrl: `https://{{HOSTNAME}}/${options.tabName}/config.html`,
+                canUpdateConfiguration: true,
+                scopes: options.tabScopes
+            });
+        }
     }
 }